Refactor: RepositoryInfoをファイルに切り出し
