+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/RepositoryInfo.kt
@@ -1,0 +1,20 @@
+package jp.co.yumemi.android.code_check
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+
+
+/**
+ * GitHubリポジトリの詳細情報を格納するクラス
+ * */
+@Parcelize
+data class RepositoryInfo(
+    val name: String,
+    val ownerIconUrl: String,
+    val language: String,
+    val stargazersCount: Int,
+    val watchersCount: Int,
+    val forksCount: Int,
+    val openIssuesCount: Int,
+) : Parcelable
