+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/SearchRepositoryViewModel.kt
@@ -3,7 +3,6 @@
  */
 package jp.co.yumemi.android.code_check
 
-import android.os.Parcelable
 import androidx.lifecycle.ViewModel
 import io.ktor.client.*
 import io.ktor.client.call.*
@@ -13,7 +12,6 @@ import io.ktor.client.statement.*
 import java.util.*
 import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
 import kotlinx.coroutines.*
-import kotlinx.parcelize.Parcelize
 import org.json.JSONObject
 
 /**
@@ -73,18 +71,3 @@ class SearchRepositoryViewModel : ViewModel() {
             }.await()
         }
 }
-
-// TODO: Modelとして単一のファイルに切り出す
-/**
- * GitHubリポジトリの詳細情報を格納するクラス
- * */
-@Parcelize
-data class RepositoryInfo(
-    val name: String,
-    val ownerIconUrl: String,
-    val language: String,
-    val stargazersCount: Int,
-    val watchersCount: Int,
-    val forksCount: Int,
-    val openIssuesCount: Int,
-) : Parcelable
